Refactor exports: concrete types in FourVectors, API in LorentzVectorBase.
Stop re-exporting kinematic accessors and separation variables from FourVectors.
Users load LorentzVectorBase for mass, pt, ΔR, etc.; FourVectors keeps
FourVector/FourVectorCyl, transforms, fromPtEtaPhiE, and fast_mass.
