+++ test/runtests.jl
@@ -1,4 +1,6 @@
 using FourVectors
+include("lorentzvectorbase_api.jl")
+__import_lorentzvectorbase!()
 using Test
 
 for (root, dirs, files) in walkdir(@__DIR__)
