+++ test/test-creation-and-interface.jl
@@ -1,4 +1,3 @@
-using FourVectors
 using Test
 
 p = FourVector(1.0, 2.0, 3.0; E = 4.0)
